vitalsgaurd/backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
phidata>=2.4.0
openai>=1.30.0          # Mistral is OpenAI-compatible; we use the Mistral endpoint
httpx>=0.27.0
numpy>=1.26.0
scikit-learn>=1.4.0
python-dotenv>=1.0.0
pydantic>=2.7.0
